更新项目包结构
