+++ app/src/main/java/com/parkingwang/vehiclekeyboard/demo/MainActivity.java
@@ -9,8 +9,8 @@ import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
 
+import com.parkingwang.vehiclekeyboard.KeyboardInputController;
 import com.parkingwang.vehiclekeyboard.core.KeyboardType;
-import com.parkingwang.vehiclekeyboard.support.KeyboardInputController;
 import com.parkingwang.vehiclekeyboard.support.PopupKeyboard;
 import com.parkingwang.vehiclekeyboard.view.InputView;
