+++ keyboard/src/main/java/com/parkingwang/vehiclekeyboard/OnNumberTypeChangedListener.java
@@ -1,0 +1,12 @@
+package com.parkingwang.vehiclekeyboard;
+
+import com.parkingwang.vehiclekeyboard.core.NumberType;
+
+/**
+ * @author 陈哈哈 (yoojiachen@gmail.com)
+ */
+public interface OnNumberTypeChangedListener {
+
+    void onChanged(NumberType currentType);
+
+}
