+++ keyboard/src/main/java/com/parkingwang/vehiclekeyboard/support/MessageHandler.java
@@ -1,8 +1,5 @@
 package com.parkingwang.vehiclekeyboard.support;
 
-public interface MessageHandler {
-
-    void onMessageError(int message);
-
-    void onMessageTip(int message);
+@Deprecated
+public interface MessageHandler extends com.parkingwang.vehiclekeyboard.MessageHandler {
 }
