+++ keyboard/src/main/java/com/parkingwang/vehiclekeyboard/support/OnInputChangedListener.java
@@ -3,9 +3,6 @@ package com.parkingwang.vehiclekeyboard.support;
 /**
  * @author 陈哈哈 (yoojiachen@gmail.com)
  */
-public interface OnInputChangedListener {
-
-    void onChanged(String number, boolean isCompleted);
-
-    void onCompleted(String number, boolean isAutoCompleted);
+@Deprecated
+public interface OnInputChangedListener extends com.parkingwang.vehiclekeyboard.OnInputChangedListener {
 }
