+++ keyboard/src/main/java/com/parkingwang/vehiclekeyboard/support/OnNumberTypeChangedListener.java
@@ -1,12 +1,8 @@
 package com.parkingwang.vehiclekeyboard.support;
 
-import com.parkingwang.vehiclekeyboard.core.NumberType;
-
 /**
  * @author 陈哈哈 (yoojiachen@gmail.com)
  */
-public interface OnNumberTypeChangedListener {
-
-    void onChanged(NumberType currentType);
-
+@Deprecated
+public interface OnNumberTypeChangedListener extends com.parkingwang.vehiclekeyboard.OnNumberTypeChangedListener {
 }
